+++ frontend/src/_services/caller.service.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const Axios = axios.create({
+    baseURL: "http://localhost:5000"
+})
+
+export default Axios;
